+++ sphinx/source/changelog.rst
@@ -19,6 +19,7 @@ Changelog
   has completed
 * Added the ``PNGScale`` configuration parameter for
   :ref:`trace.py <trace-conf>` (to specify the scale factor of the PNG image)
+* :ref:`snapinfo.py` now shows AY register values in 128K SZX and Z80 snapshots
 * Added support to the :ref:`FOREACH` macro for the ``POKEname`` special
   variable
 * Added support to the :ref:`LET` macro for setting individual key-value pairs